move default activation code to its own package
